requirements.txt
flake8==3.8.3
numpy==1.18.5
pillow==7.1.2

